+++ src/main/java/com/prgrms/boardjpa/commons/api/SuccessResponse.java
@@ -1,4 +1,4 @@
-package com.prgrms.boardjpa.commons;
+package com.prgrms.boardjpa.commons.api;
 
 public class SuccessResponse<T> {
 	private final T data;
