+++ src/test/java/com/prgrms/boardjpa/post/PostRestControllerTest.java
@@ -25,7 +25,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.prgrms.boardjpa.commons.SuccessResponse;
+import com.prgrms.boardjpa.commons.api.SuccessResponse;
 import com.prgrms.boardjpa.post.dto.PostDto;
 import com.prgrms.boardjpa.user.domain.Hobby;
 import com.prgrms.boardjpa.user.domain.User;
